grades

+++ ServerImpl/communication/Controllers/AnswerQuestionController.cs
@@ -100,7 +100,7 @@ namespace communication.Controllers
             }
             else if (ans.Equals(Replies.SUCCESS))
             {
-                int grade = ServerWiring.getInstance().getTestGrade(Convert.ToInt32(userCookie.Value), groupCookie.Value, Convert.ToInt32(testCookie.Value));
+                double grade = ServerWiring.getInstance().getTestGrade(Convert.ToInt32(userCookie.Value), groupCookie.Value, Convert.ToInt32(testCookie.Value));
                 return RedirectToAction("Index", "Main", new { message = "your grade is "+grade});
             }
             else
